inventory_srv/requirements.txt
peewee
pymysql
passlib
grpcio
grpcio-tools
protobuf
cryptography
loguru
python-consul2
nacos-sdk-python